Force n_fft cast from float to int

+++ magenta/models/nsynth/spektra.py
@@ -19,7 +19,7 @@ frameShiftMS = 10
 # parameter setup
 mag = mgu.specgram(audio)  # y?
 phase_angle = 0
-n_fft = pow(2, (math.ceil(math.log(frameLen)/math.log(2))))
+n_fft = pow(2, int(math.ceil(math.log(frameLen)/math.log(2))))
 hop = frameShiftMS
 num_iters = 50
